added another endpoint to EF controllers

+++ Models/UserSalary.cs
@@ -3,6 +3,6 @@ namespace DotnetAPI.Models
     public partial class UserSalary
     {
         public int UserId { get; set; }
-        public float Salary { get; set; }
+        public double Salary { get; set; }
     }
 }
